Updates for DR18

+++ Sciserver_webService/ToolsSearch/ObjectInfo.cs
@@ -17,10 +17,10 @@ namespace Sciserver_webService.ToolsSearch
         public double? ra = null;
         public double? dec = null;
 
-        public int? mjd = null;
-        public short? plate = null;
+        public long? mjd = null;
+        public decimal? plate = null;
         public decimal? plateId = null;
-        public short? fiberId = null;
+        public long? fiberId = null;
         public string fieldId = null;
 
         public int? run = null;
